create basecollection view
