+++ frontend/src/components/common/Collection/RAvatar.vue
@@ -1,9 +1,5 @@
 <script setup lang="ts">
-import type {
-  Collection,
-  VirtualCollection,
-  SmartCollection,
-} from "@/stores/collections";
+import { type CollectionType } from "@/stores/collections";
 import storeCollections from "@/stores/collections";
 import { getCollectionCoverImage, getFavoriteCoverImage } from "@/utils/covers";
 import { computed, ref, watchEffect } from "vue";
@@ -11,7 +7,7 @@ import { useTheme } from "vuetify";
 
 const props = withDefaults(
   defineProps<{
-    collection: Collection | VirtualCollection | SmartCollection;
+    collection: CollectionType;
     size?: number;
   }>(),
   {
